refactor: login page 로직과 UI 분리
